New energy config parsing algorithm has been used

+++ Efficiency/include/bdt.h
@@ -162,7 +162,7 @@ public:
 		const std::vector<double> &fraclayer,
 		const double fraclastlayer,
 		const double corrected_energy_gev,
-		const std::vector<float> &energy_binning,
+		const std::vector<double> &energy_binning,
 		const TVector3& bgo_direction);
 
     const double GetBDTCut_10_100();
